Add ability add members to a postgres role
